change js file into modules
